refactor: 승패무 판단을 Result의 책임으로 이동 (재미있게 BiPredicate 활용)
