+++ src/main/java/domain/Result.java
@@ -1,15 +1,34 @@
 package domain;
 
+import java.util.Arrays;
+import java.util.function.BiPredicate;
+
 public enum Result {
 
-    WIN("승"),
-    LOSE("패"),
-    TIE("무");
+    WIN("승", Result::winningCondition),
+    TIE("무", Result::tieCondition),
+    LOSE("패", Result::loseCondition);
 
     private final String value;
+    private final BiPredicate<Hands, Hands> condition;
 
-    Result(final String value) {
+    Result(final String value, final BiPredicate<Hands, Hands> condition) {
         this.value = value;
+        this.condition = condition;
+    }
+
+    private static boolean winningCondition(final Hands hands, final Hands target) {
+        return (!hands.isBust() && target.isBust())
+                || (hands.sum() > target.sum() && !hands.isBust())
+                || (hands.sum() == target.sum() && hands.size() < target.size() && !hands.isBust());
+    }
+
+    private static boolean tieCondition(final Hands hands, final Hands target) {
+        return hands.sum() == target.sum() && hands.size() == target.size() && !hands.isBust();
+    }
+
+    private static boolean loseCondition(final Hands hands, final Hands target) {
+        return hands.isBust() || hands.sum() < target.sum() || !target.isBust();
     }
 
     public Result reverse() {
@@ -20,6 +39,13 @@ public enum Result {
             return WIN;
         }
         return TIE;
+    }
+
+    public static Result calculateOf(final Hands hands, final Hands target) {
+        return Arrays.stream(Result.values())
+                .filter(result -> result.condition.test(hands, target))
+                .findFirst()
+                .orElseThrow();
     }
 
     public String getValue() {
